change the directory name
