--- model/conf/requirements.txt
@@ -1,13 +1,0 @@
-Keras==2.4.3
-Keras-Preprocessing==1.1.2
-librosa==0.7.0
-numba==0.48.0
-numpy==1.16.6
-pandas==1.2.0
-pydub==0.24.1
-PyYAML==5.3.1
-scikit-learn==0.24.0
-scipy==1.6.0
-urllib3==1.22
-soundfile
-
