--- model/saved_models/saved_model.h5
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e36ab69fb91ca7a41e83e715446e749028e084c260665c244c3b29d7e76b1b5c
-size 6749016
